requirements.txt
anyio==3.6.2
certifi==2022.12.7
contourpy==1.0.6
cycler==0.11.0
emoji==2.2.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
isOdd==0.1.2
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.0b0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0